Added few funtionality to the application
When user logged in- the system checks whether the user authenticated his/her main profile to send emails from this application. if the user authenticated before, then the user can send email. if not, the dashboard shows the process to how to authenticate before sending emails. (valid only for gmail accounts).